+++ MGen/GenMP1.h
@@ -8,7 +8,7 @@ public:
 	CGenMP1();
 	~CGenMP1();
 	void Generate() override;
-	void LoadConfigLine(CString * sN, CString * sV, int idata, double fdata);
+	void LoadConfigLine(CString * sN, CString * sV, int idata, float fdata);
 
 	CString midi_file;
 };
